Allow alias to be used with imsi tracking file.
The FUT project end users want to be able to but a recognizable string
next to the imsis in the tracking file. This way they can relate an Imsi
to a specific person. POST of the string is already there, format communicated
to end users will be "imsi(alias)". Update DELETE endpoint to be able to
still delete an imsi by the number so needed to filter out the alias
portion when evaluating whether the to delete matches an imsi in the
tracking file.