Use config feature of Flask framework.
Simplify configuration of Flask framework to all for multiple environments to be defined and supported (i.e. DEV, QA, and PROD) with their own respective settings.
This change allows for easy development and deployment configuration settings to be controlled at the environment variable level as well as defaults when not specified.