Allow environment variables and values to be read in from json config file and set as class instances attributes dynamically from json data converted to python dictionary. Also add in logger json config file template.
Environment variables from json config file allows for all environment variables to be read in, configured, and set at a class object level as opposed to be individually accessed from the OS environ dictionary. Allows for different execution environments, test, QA, Prod, etc, to be set differently without having internally code for the existence or non existence of a key with or without a value depending on execution environments interpretation of OS environment variables.